example no.2 (ETS)
We set a CO2 emissions cap (threshold).
We allow emissions to exceed this cap, but any excess is tracked.
The excess emissions are 'penalized' with an additional cost in the objective function. This is a simple, realistic and effective approach to represent the actual EU Emissions Trading System (ETS) regulation where companies can exceed their caps by purchasing allowances.